+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Remove initial memory argument [#727]
 - Add name of platform an instance is from when launching [#740]
 - Add in Get Help button to more instances when available [#734]
+- Add dialog warning user when trying to skip external download mods from CurseForge
 
 ### Fixes
 - Remove old OmitStackTraceInFastThrow JVM arg